take args

+++ __main__.py
@@ -3,6 +3,7 @@ sp6 merge
 """
 import os
 import sys
+import getopt
 import configparser
 import traceback
 import time
@@ -11,9 +12,11 @@ import hashcalc
 VERSION = 'V0.1'
 DATE = '20180601'
 
-WORKING_PATH = os.path.split(os.path.abspath(sys.argv[0]))[0]
-SORTWARE_PATH = sys._MEIPASS if getattr(sys, 'frozen', False) else WORKING_PATH
-CONFIG_FILE = os.path.join(WORKING_PATH, r'appupdate.ini')
+TEMP_PATH = os.path.join(os.environ["TMP"], str(time.time()).split('.')[0])
+WORKING_PATH = None
+SOFTWARE_PATH = os.path.join(os.path.split(os.path.abspath(sys.argv[0]))[0])
+SOFTWARE_REAL_PATH = sys._MEIPASS if getattr(sys, 'frozen', False) else SOFTWARE_PATH
+CONFIG_FILE = os.path.join(SOFTWARE_PATH, r'appupdate.ini')
 CONFIG_FILE_CONTENT_DEFAULT =\
 r'''[outfile]
 path = update.img
@@ -69,14 +72,9 @@ class ConfigClass():
 CONFIG = ConfigClass()
 
 
-def main():
+def create_img(outpath):
     """main"""
     try:
-        CONFIG.chk_config()
-        out_file_path = os.path.join(WORKING_PATH, '..', CONFIG.outfile_cfg().get('path'))
-        out_file_path = os.path.abspath(out_file_path)
-        print('out path:', out_file_path)
-
         sh_file_path = os.path.join(WORKING_PATH, 'autorun.sh')
         md5_file_path = os.path.join(WORKING_PATH, 'bin.md5')
         if os.path.isfile(sh_file_path):
@@ -84,71 +82,77 @@ def main():
         if os.path.isfile(md5_file_path):
             os.remove(md5_file_path)
 
-        dos2unix_path = os.path.join(SORTWARE_PATH, r'mkcramfs\dos2unix.exe')
+        dos2unix_path = os.path.join(SOFTWARE_REAL_PATH, r'mkcramfs\dos2unix.exe')
 
         # write script
         files = ['%s'%x for x in os.listdir(WORKING_PATH) if os.path.isfile(os.path.join(WORKING_PATH, x))]
         sh_content = UPDATE_SCRIPT_FORMAT%(' '.join(files))
         with open(sh_file_path, 'w') as sh_file:
             sh_file.write(sh_content)
-        os.system(dos2unix_path + ' ' + sh_file_path)
+        os.system(dos2unix_path + ' "' + sh_file_path + '"')
 
         # calc md5
         md5_content = hashcalc.md5sum(WORKING_PATH)
         with open(md5_file_path, 'w') as md5_file:
             md5_file.write(md5_content)
-        os.system(dos2unix_path + ' ' + md5_file_path)
+        os.system(dos2unix_path + ' "' + md5_file_path + '"')
 
         # mk cramfs
-        mkcramfs_path = os.path.join(SORTWARE_PATH, r'mkcramfs\mkcramfs.exe')
-        cmd = mkcramfs_path + ' ' + WORKING_PATH + ' ' + os.path.basename(CONFIG.outfile_cfg().get('path'))
+        mkcramfs_path = os.path.join(SOFTWARE_REAL_PATH, r'mkcramfs\mkcramfs.exe')
+        cmd = mkcramfs_path + ' "' + WORKING_PATH + '" "' + outpath + '"'
         print('cmd:', cmd)
         if os.system(cmd) != 0:
             raise Exception('mkcramfs error, merge abort.')
-
-        # del sh and md5
-        os.remove(sh_file_path)
-        os.remove(md5_file_path)
         return 0
     except Exception:
         traceback.print_exc()
         return -1
+    finally:
+        # del sh
+        if os.path.isfile(sh_file_path):os.remove(sh_file_path)
+        # if os.path.isfile(md5_file_path):os.remove(md5_file_path)
 
 
-def del_outfile():
+def del_outfile(outpath):
     """delete outfile"""
     try:
-        out_merge_file_path = os.path.join(CONFIG.outfile_cfg().get('path'))
-        if os.path.isfile(out_merge_file_path):
-            os.remove(out_merge_file_path)
+        if os.path.isfile(outpath):
+            os.remove(outpath)
     except Exception:
         traceback.print_exc()
         print('outfile del failed.')
 
 if __name__ == '__main__':
-    if len(sys.argv) < 2:
-        print('ERROR: 请将需要打包的文件夹拖到本软件上进行打包')
+    try:
+        CONFIG.chk_config()
+        opts, args = getopt.getopt(sys.argv[1:], "o:")
+        if not args:
+            raise Exception('ERROR: 请将需要打包的文件夹拖到本软件上进行打包')
+        WORKING_PATH = args[0]
+        out_file_path = os.path.join(WORKING_PATH, '..', CONFIG.outfile_cfg().get('path'))
+        for op, value in opts:
+            if op == '-o':
+                out_file_path = os.path.join(value)
+        if not os.path.isdir(WORKING_PATH):
+            raise Exception('ERROR: working path{path} invalid.'.format(path=WORKING_PATH))
+
+        tm_start = time.time()
+        print('SP6 App Update Creator {ver}({date}).Designed by Kay.'.format(ver=VERSION, date=DATE))
+        print('WORKING_PATH:', WORKING_PATH)
+        print('CONFIG_FILE:', CONFIG_FILE)
+        print('out file:', out_file_path)
+
+        if create_img(out_file_path) == 0:
+            print('success')
+        else:
+            raise Exception('!!FAILED!!')
+        print('time use {tm:.1f}s'.format(tm=time.time() - tm_start))
+        sys.exit(0)
+    except Exception:
+        traceback.print_exc()
         os.system('color 47')
+        del_outfile(out_file_path)
         time.sleep(3)
         os.system('color 07')
-        sys.exit(1)
-    WORKING_PATH = os.path.abspath(sys.argv[1])
-    if not os.path.isdir(WORKING_PATH):
-        print('ERROR: working path invalid.')
         sys.exit(1)
 
-    tm_start = time.time()
-    print('SP6 App Update Creator {ver}({date}).Designed by Kay.'.format(ver=VERSION, date=DATE))
-    print('WORKING_PATH:', WORKING_PATH)
-    print('CONFIG_FILE:', CONFIG_FILE)
-    if main() == 0:
-        print('success')
-    else:
-        print('!!FAILED!!')
-        os.system('color 47')
-        del_outfile()
-        time.sleep(3)
-        os.system('color 07')
-        sys.exit(1)
-    print('time use {tm:.1f}s'.format(tm=time.time() - tm_start))
-    sys.exit(0)
